requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dacite==1.9.2
Deprecated==1.2.18
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-flask==0.47b0
prometheus-client==0.20.0
Flask==3.1.1
flask-cors==6.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
jaconv==0.4.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mutagen==1.47.0
platformdirs==4.3.8
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pykakasi==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytube==15.0.0
RapidFuzz==3.13.0
redis==6.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soundcloud-v2==1.6.0
soupsieve==2.7
spotdl==4.4.2
spotipy==2.25.1
SQLAlchemy==2.0.41
syncedlyrics==1.0.1
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.13.2
Unidecode==1.4.0
urllib3==2.4.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
yt-dlp==2025.08.27,<2026
ytmusicapi==1.11.1,<2
pydub==0.25.1
comtypes==1.4.6
